requirements.txt
# data preprocessing
pdfplumber==0.10.3
python-docx==1.1.0

# training
easydict==1.11
torch==2.1.2
transformers==4.34.1
bitsandbytes==0.41.1
scipy==1.11.3
sentencepiece==0.1.99
protobuf==4.25.1
wandb==0.15.12
datasets==2.5.2
git+https://github.com/huggingface/peft.git  # LoftQ

# others
gdown==4.7.1

# for conveniently run python scripts
ipython
